pyproject.toml
[tool.poetry]
name = "aid_coordinator"
version = "0.1.0"
description = ""
authors = ["Sander Steffann <sander@steffann.nl>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.3"
daphne = "^3.0.2"
gunicorn = "^20.1.0"
django-debug-toolbar = "^3.2.4"
django-xff = "^1.3.0"
mysqlclient = "^2.1.0"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-filter = "^21.1"
Jinja2 = "^3.0.3"
Pillow = "^9.0.1"
django-import-export = "^2.8.0"
django-admin-wizard = "^0.3.0"
django-registration = "^3.2"
django-countries = {extras = ["pyuca"], version = "^7.3.2"}
django-phonenumber-field = {extras = ["phonenumbers"], version = "^6.3.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
